Refactor files and multipart
