+++ src/Organisation/OrganisationFile.php
@@ -6,47 +6,8 @@
 
 namespace Wizaplace\SDK\Organisation;
 
-use Psr\Http\Message\StreamInterface;
+use Wizaplace\SDK\File\FileService;
 
-class OrganisationFile
+class OrganisationFile extends FileService
 {
-    /** @var string */
-    private $name;
-
-    /** @var StreamInterface */
-    private $contents;
-
-    /** @var string */
-    private $filename;
-
-    public function __construct(string $name, StreamInterface $contents, string $filename)
-    {
-        $this->name = $name;
-        $this->contents = $contents;
-        $this->filename = $filename;
-    }
-
-    /**
-     * @return string
-     */
-    public function getName(): string
-    {
-        return $this->name;
-    }
-
-    /**
-     * @return StreamInterface
-     */
-    public function getContents(): StreamInterface
-    {
-        return $this->contents;
-    }
-
-    /**
-     * @return string
-     */
-    public function getFilename(): string
-    {
-        return $this->filename;
-    }
 }
