+++ src/Pim/MultiVendorProduct/MultiVendorProductService.php
@@ -10,9 +10,9 @@ namespace Wizaplace\SDK\Pim\MultiVendorProduct;
 use GuzzleHttp\Exception\ClientException;
 use GuzzleHttp\RequestOptions;
 use Wizaplace\SDK\AbstractService;
-use Wizaplace\SDK\ArrayableInterface;
 use Wizaplace\SDK\Exception\NotFound;
 use Wizaplace\SDK\Exception\SomeParametersAreInvalid;
+use Wizaplace\SDK\File\FileService;
 use function theodorejb\polycast\to_string;
 
 final class MultiVendorProductService extends AbstractService
@@ -72,7 +72,7 @@ final class MultiVendorProductService extends AbstractService
 
         try {
             $response = $this->client->post("pim/multi-vendor-products/{$mvpId}/images", [
-                RequestOptions::MULTIPART => $this->createMultipartArray([], $files),
+                RequestOptions::MULTIPART => FileService::createMultipartArray([], $files),
             ]);
         } catch (ClientException $e) {
             if ($e->getCode() === 404) {
@@ -85,72 +85,5 @@ final class MultiVendorProductService extends AbstractService
         }
 
         return new MultiVendorProduct($response);
-    }
-
-    /**
-     * This method help to have an array compliant to Guzzle for multipart POST/PUT for the organisation process
-     * There are exception in the process for OrganisationAddress and OrganisationAdministrator which needs to be transformed to array
-     * prior to processing
-     *
-     * Ex:
-     * ['name' => 'obiwan', ['address' => ['street' => 'main street', 'city' => 'Mos Esley']]
-     * needs to be flatten to
-     * ['name' => 'obiwan', 'address[street]' => 'main street', 'address[city]' => 'Mos esley']
-     *
-     * @param array $array
-     * @param string $originalKey
-     * @return array
-     */
-    private function flattenArray(array $array, string $originalKey = '')
-    {
-        $output = [];
-
-        foreach ($array as $key => $value) {
-            $newKey = $originalKey;
-            if (empty($originalKey)) {
-                $newKey .= $key;
-            } else {
-                $newKey .= '['.$key.']';
-            }
-
-            if (is_array($value)) {
-                $output = array_merge($output, $this->flattenArray($value, $newKey));
-            } elseif ($value instanceof ArrayableInterface) {
-                $output = array_merge($output, $this->flattenArray($value->toArray(), $newKey));
-            } else {
-                $output[$newKey] = $value;
-            }
-        }
-
-        return $output;
-    }
-
-    /**
-     * @param array                    $data
-     * @param MultiVendorProductFile[] $files
-     *
-     * @return array
-     */
-    private function createMultipartArray(array $data, array $files) : array
-    {
-        $dataToSend = [];
-
-        $flatArray = $this->flattenArray($data);
-
-        foreach ($flatArray as $key => $value) {
-            $dataToSend[] = [
-                'name'  => $key,
-                'contents' => $value,
-            ];
-        }
-
-        foreach ($files as $file) {
-            $dataToSend[] = [
-                'name' => $file->getName(),
-                'contents' => $file->getContents(),
-            ];
-        }
-
-        return $dataToSend;
     }
 }
