+++ tests/Company/CompanyServiceTest.php
@@ -8,8 +8,6 @@ declare(strict_types = 1);
 namespace Wizaplace\SDK\Tests\Company;
 
 use GuzzleHttp\Exception\ClientException;
-use PHPUnit_Framework_MockObject_MockObject;
-use Psr\Http\Message\UploadedFileInterface;
 use Wizaplace\SDK\Authentication\AuthenticationRequired;
 use Wizaplace\SDK\Company\Company;
 use Wizaplace\SDK\Company\CompanyRegistration;
@@ -19,7 +17,7 @@ use Wizaplace\SDK\Company\CompanyUpdateCommand;
 use Wizaplace\SDK\Company\UnauthenticatedCompanyRegistration;
 use Wizaplace\SDK\Exception\CompanyNotFound;
 use Wizaplace\SDK\Tests\ApiTestCase;
-use function GuzzleHttp\Psr7\stream_for;
+use Wizaplace\SDK\Tests\File\FileTestService;
 
 /**
  * @see CompanyService
@@ -45,8 +43,8 @@ final class CompanyServiceTest extends ApiTestCase
         $companyRegistration->setUrl('https://acme.example.com/');
         $companyRegistration->setExtra(['driving_license_number' => '654987321']);
 
-        $companyRegistration->addUploadedFile('rib', $this->mockUploadedFile('minimal.pdf'));
-        $companyRegistration->addUploadedFile('idCard', $this->mockUploadedFile('minimal.pdf'));
+        $companyRegistration->addUploadedFile('rib', FileTestService::mockUploadedFile('minimal.pdf'));
+        $companyRegistration->addUploadedFile('idCard', FileTestService::mockUploadedFile('minimal.pdf'));
 
         $companyService = $this->buildUserCompanyService('customer-3@world-company.com', 'password-customer-3');
 
@@ -87,7 +85,7 @@ final class CompanyServiceTest extends ApiTestCase
 
 
         // Update file
-        $file = $this->mockUploadedFile('minimal.pdf');
+        $file = FileTestService::mockUploadedFile('minimal.pdf');
 
         $update = $companyService->updateFile($company->getId(), 'idCard', [
             'name'     => "idCard",
@@ -147,7 +145,7 @@ final class CompanyServiceTest extends ApiTestCase
     public function testUploadingBadExtensionRegistrationFiles()
     {
         $companyRegistration = new CompanyRegistration('4CME Test Inc', 'acme4@example.com');
-        $companyRegistration->addUploadedFile('rib', $this->mockUploadedFile('dummy.txt'));
+        $companyRegistration->addUploadedFile('rib', FileTestService::mockUploadedFile('dummy.txt'));
         $companyService = $this->buildUserCompanyService('customer-3@world-company.com', 'password-customer-3');
 
         $result = $companyService->register($companyRegistration);
@@ -245,17 +243,5 @@ final class CompanyServiceTest extends ApiTestCase
         $apiClient->authenticate($email, $password);
 
         return new CompanyService($apiClient);
-    }
-
-    private function mockUploadedFile(string $filename): UploadedFileInterface
-    {
-        $path = __DIR__.'/../fixtures/files/'.$filename;
-
-        /** @var UploadedFileInterface|PHPUnit_Framework_MockObject_MockObject $file */
-        $file = $this->createMock(UploadedFileInterface::class);
-        $file->expects($this->once())->method('getStream')->willReturn(stream_for(fopen($path, 'r')));
-        $file->expects($this->once())->method('getClientFilename')->willReturn($filename);
-
-        return $file;
     }
 }
